app/requirements.txt
PyQt6>=6.5.0
requests>=2.31.0
python-docx>=1.0.0
pysrt>=1.1.2
langdetect>=1.0.9
PySocks>=1.7.1
aiohttp>=3.9.0
pydub>=0.25.1
pytest>=7.4.0
pytest-cov>=4.1.0
pyinstaller>=6.0.0
pillow>=10.0.0
supabase>=2.0.0
qasync>=0.27.0
numpy>=1.24.0
packaging>=23.0
